Move ResourceSchema to the 'connector' module
